Fixed bug uploading files with x-amz-acl header to S3.
Uploads to S3 which use the x-amz-acl header were failing with an
authentication error. When this header was included in the canonical
headers uploads worked OK.

Reworked AWS4Auth.get_canonical_headers() to include by default
all headers starting in x-amz-. Changed default included non-x-
amz- headers to host and content-type, to match AWS docs.

On testing found that mobileanalytics auth failed when including
the x-amz-client-context header in canonical headers, so added
an exlusion for this header to get_canonical_headers().